Update: Restructure API object

+++ irc_bot/src/osu_api/osu_api_structures/auth_request.rs
@@ -2,7 +2,7 @@ use serde::Serialize;
 
 #[derive(Serialize)]
 pub struct AuthRequest {
-    pub client_id: i64,
+    pub client_id: i32,
     pub client_secret: String,
     pub grant_type: String,
     pub scope: String,
